feat(parameters helper): moves getCustomMetaDataWithIpfsCid into it PE-1886
